Refactor into separate classes for ftp and object storage and coordinator
